+++ assets/js/blocks/product-collection/index.tsx
@@ -12,7 +12,7 @@ import edit from './edit';
 import save from './save';
 import icon from './icon';
 import './variations';
-import { getDefaultStockStatuses } from './inspector-controls/utils';
+import { getDefaultStockStatuses } from './inspector-controls/constants';
 
 if ( isExperimentalBuild() ) {
 	registerBlockType( metadata, {
