+++ assets/js/blocks/product-collection/inspector-controls/constants.ts
@@ -1,4 +1,10 @@
 /**
+ * External dependencies
+ */
+import { getSetting } from '@woocommerce/settings';
+import { objectOmit } from '@woocommerce/utils';
+
+/**
  * Internal dependencies
  */
 import blockJson from '../block.json';
@@ -10,8 +16,25 @@ import {
 
 const defaultQuery = blockJson.attributes.query.default;
 
+export const STOCK_STATUS_OPTIONS = getSetting< Record< string, string > >(
+	'stockStatusOptions',
+	[]
+);
+
+const GLOBAL_HIDE_OUT_OF_STOCK = getSetting< boolean >(
+	'hideOutOfStockItems',
+	false
+);
+
+export const getDefaultStockStatuses = () => {
+	return GLOBAL_HIDE_OUT_OF_STOCK
+		? Object.keys( objectOmit( STOCK_STATUS_OPTIONS, 'outofstock' ) )
+		: Object.keys( STOCK_STATUS_OPTIONS );
+};
+
 export const DEFAULT_FILTERS = {
 	woocommerceOnSale: defaultQuery.woocommerceOnSale,
+	woocommerceStockStatus: getDefaultStockStatuses(),
 };
 
 export const getDefaultSettings = (
